update closure templates so it works with updated closure library

+++ download-libs.sh
@@ -84,8 +84,11 @@ fi
 # Soy file bundled with the compiler does not compile with strict settings:
 # lib/closure-templates-compiler/soyutils_usegoog.js:1762: ERROR - element JS_STR_CHARS does not exist on this enum
 cd closure-templates-compiler
-curl https://raw.githubusercontent.com/google/closure-templates/0cbc8543c34d3f7727dd83a2d1938672f16d5c20/javascript/soyutils_usegoog.js -O
-mv soyutils.js soyutils.js.notused
+curl https://raw.githubusercontent.com/google/closure-templates/05890ed973229bb8ae8a718b01b6ad80560243eb/javascript/soyutils_usegoog.js -O
+if [ -f soyutils.js ]; then
+  # soyutuls.js re-declares goog, which breaks closure
+  mv soyutils.js soyutils.js.notused
+fi
 cd ..
 
 cd ..
